+++ public/index.html
@@ -46,6 +46,7 @@
 
       <!-- Контент -->
     </main>
-    <script src="../src/main.js"></script>
+    <script src="bundle.js"></script>
+    <!-- <script src="../src/main.js"></script> -->
   </body>
 </html>
