+++ src/components/load-more.js
@@ -1,0 +1,7 @@
+export const getLoadMoreComponent = () => {
+  const markup = `
+    <button class="load-more" type="button">load more</button>
+  `;
+  // return getComponent(markup);
+  return markup;
+};
