+++ src/components/render.js
@@ -1,0 +1,3 @@
+export const renderComponent = (container, component, place) => {
+  container.insertAdjacentHTML(place, component);
+};
